Authorization, Viewsets and URL factorization

+++ users/api.py
@@ -7,40 +7,45 @@ from rest_framework.pagination import PageNumberPagination
 from rest_framework.renderers import JSONRenderer
 from rest_framework.response import Response
 from rest_framework.views import APIView
-
+from users.permissions import UserPermission
 from users.serializers import UserSerializer
+from rest_framework.viewsets import ViewSet
 
+class UserViewSet(ViewSet):
+    permission_classes = (UserPermission,)
 
-class UserListAPI(APIView):
-
-    def get(self, request):
+    def list(self, request):
+        self.check_permissions(request)
         paginator = PageNumberPagination()
         users = User.objects.all()
 
-        paginator.paginate_queryset(users,request)
+        paginator.paginate_queryset(users, request)
         serializer = UserSerializer(users, many=True)
-        serialized_users = serializer.data #dictionaries list
+        serialized_users = serializer.data  # dictionaries list
 
         return paginator.get_paginated_response(serialized_users)
 
-    def post(self, request):
+    def create(self, request):
+        self.check_permissions(request)
         serializer = UserSerializer(data=request.data)
         if serializer.is_valid():
             new_user = serializer.save()
-            return Response(serializer.data, status = status.HTTP_201_CREATED)
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
         else:
-            return Response(serializer.errors, status = status.HTTP_400_BAD_REQUEST)
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-class UserDetailAPI(APIView):
-
-    def get(self, request, pk):
+    def retrieve(self, request, pk):
+        self.check_permissions(request)
         user = get_object_or_404(User, pk=pk)
+        self.check_object_permissions(request,user)
         serializer = UserSerializer(user)
         serialized_user = serializer.data
         return Response(serialized_user)
 
-    def put(self, request,pk):
+    def update(self, request, pk):
+        self.check_permissions(request)
         user = get_object_or_404(User, pk=pk)
+        self.check_object_permissions(request, user)
         serializer = UserSerializer(instance=user, data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -48,7 +53,64 @@ class UserDetailAPI(APIView):
         else:
             return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-    def delete(self,request,pk):
+    def destroy(self, request, pk):
+        self.check_permissions(request)
         user = get_object_or_404(User, pk=pk)
+        self.check_object_permissions(request, user)
         user.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
+
+"""
+class UserListAPI(APIView):
+    permission_classes = (UserPermission,)
+
+    def get(self, request):
+        self.check_permissions(request)
+        paginator = PageNumberPagination()
+        users = User.objects.all()
+
+        paginator.paginate_queryset(users, request)
+        serializer = UserSerializer(users, many=True)
+        serialized_users = serializer.data  # dictionaries list
+
+        return paginator.get_paginated_response(serialized_users)
+
+    def post(self, request):
+        self.check_permissions(request)
+        serializer = UserSerializer(data=request.data)
+        if serializer.is_valid():
+            new_user = serializer.save()
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        else:
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+
+class UserDetailAPI(APIView):
+
+    permission_classes = (UserPermission,)
+
+    def get(self, request, pk):
+        self.check_permissions(request)
+        user = get_object_or_404(User, pk=pk)
+        self.check_object_permissions(request,user)
+        serializer = UserSerializer(user)
+        serialized_user = serializer.data
+        return Response(serialized_user)
+
+    def put(self, request, pk):
+        self.check_permissions(request)
+        user = get_object_or_404(User, pk=pk)
+        self.check_object_permissions(request, user)
+        serializer = UserSerializer(instance=user, data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data, status=status.HTTP_200_OK)
+        else:
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+    def delete(self, request, pk):
+        self.check_permissions(request)
+        user = get_object_or_404(User, pk=pk)
+        self.check_object_permissions(request, user)
+        user.delete()
+        return Response(status=status.HTTP_204_NO_CONTENT)"""
